# Speed up dominators algorithm