--- a/rir/src/compiler/analysis/liveness.cpp
+++ b/rir/src/compiler/analysis/liveness.cpp
@@ -150,7 +150,7 @@ restart:
     }
 
     Visitor::run(code->entry, [&](BB* bb) {
-        for (auto n : bb->succsessors()) {
+        for (auto n : bb->successors()) {
             if (!liveAtEnd.count(n))
                 todo.insert(n);
         }