I rewrote the dominators computation to use the Lenguaer-Tarjan algorithm. There were a few drive-by commits I included:

- Typo in BB successors method.
- LT algorithm assumes the start node has no predecessors, so I made some changes to check/enforce that.
- Previously, the `dominates` method only checked for strict dominance, i.e. a node does not dominate itself. I made `dominates` check for non-strict dominance and added a `strictlyDominates` method. I also went and changed other places in the codebase to use `strictlyDominates` if it made sense. @skrynski reviewed that commit for me.

The implementation is thoroughly documented, but it is a fairly close translation of the pseudocode from [_Modern Compiler Implementation_](https://www.cs.princeton.edu/~appel/modern/).

If you have any branches that use dominance, please double-check after merging in this branch. Check whether you really want strict or non-strict dominanace.

### Performance numbers

On the flexclust benchmark, this speeds up the compiler overhead by about 2 seconds. I ran flexclust with our benchmark framework, but also inserted timing/counters into the dominators code. These timers/counters are across all runs of the benchmark.

**Before this PR**
```
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 16346475us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 14905355us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 21660213us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 21436819us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 14856454us
flexclust_no_s4: iterations=5; average: 17841063 us; total: 89205316us
=== DominanceGraph perf:
                DomGraph ctor    4988 ms (called     61940 times)
               DomGraph merge    4434 ms (called   1732656 times)
          DomGraph dominates?      42 ms (called    911066 times)
         DomGraph idominates?     186 ms (called   7798074 times)
```

We can see that there is some overhead on the first run, and significant overhead on the third and fourth runs. The time for `DomGraph merge` is included inside `DomGraph ctor`. Note that while we spend 5 seconds inside `DomGraph ctor`, that time is mainly split over the third and fourth runs, i.e. we spend 2.5 seconds constructing the dominance graph.

**With this PR**
```
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 15604257us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 14444395us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 18667478us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 18996542us
Flexclust::kcca  5  352.0658   74
flexclust_no_s4: iterations=1 runtime: 14712814us
flexclust_no_s4: iterations=5; average: 16485097 us; total: 82425486us
=== DominanceGraph perf:
                DomGraph ctor      90 ms (called     61950 times)
          DomGraph dominates?     169 ms (called    928404 times)
         DomGraph idominates?     153 ms (called   8083418 times)
```

There is a similar pattern for when we see the overheads. However, the time for the slow runs (numbers three and four) is less, about 2 seconds faster than before. We also see that we spend significantly less time inside `DomGraph ctor`.

`DomGraph dominates` got a bit slower -- before, it was looking up a value in an array. But now, we need to walk up the dominator tree. We could precompute this, but I don't think it's worth the trouble.