@mhyee 2020-04-14
> most beautiful PR award 😂
> thanks

Haha, this was the only way I had a chance of understanding what the algorithm was doing. I think I understand _what_ it's doing, but _how_ and _why_ are different matters...

Also, completely unscientific "benchmarking": the CI run for this merge took 4 hours, but previous runs would take 4.5 to 4.75 hours.